+++ src/test_engine.h
@@ -7,4 +7,4 @@ void set_screen_size(float width, float height);
 
 void update_screen(void);
 
-void on_touch(int id, float x, float y, int event);
+void on_touch(unsigned long long id, float x, float y, int event);
